rename : store 폴더명 변경에 따른 import 경로 수정
