+++ src/pages/My-Page/MyPage.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useUserStore } from '../../stores/store';
+import { useUserStore } from '../../store/store';
 import api from '../../services/axios';
 
 import Password from './PasswordChange';
