+++ src/pages/PostDetail.tsx
@@ -1,16 +1,23 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-import { Post } from '../stores/post-store';
+import { Post } from '../store/post-store';
 import api from '../services/axios';
+import { useUserStore } from '../store/store';
 
 const fetchPost = async (postId: string) => {
     const res = await api.get(`/posts/${postId}`);
     return res.data;
   };
 
-const toggleLike = async (postId: string) => {
-  const res = await api.post(`/posts/${postId}/like`);
+const toggleLike = async ({postId, token} : { postId: string; token: string }) => {
+  const res = await api.post(
+    `/posts/${postId}/like`, {},
+    {
+      headers : {
+        Authorization: `Bearer ${token}`,
+      },
+    });
   return res.data;
 };
 
@@ -18,6 +25,8 @@ const PostDetail = () => {
    
   const { postId } = useParams<{ postId: string }>();
   const queryClient = useQueryClient()
+  const token = useUserStore.getState().accessToken;
+
 
   const { data: post, isLoading, isError } = useQuery<Post>({
     queryKey: ['postDetail', postId],
@@ -26,9 +35,12 @@ const PostDetail = () => {
   });
 
   const { mutate: likePost, status: likeStatus } = useMutation({
-    mutationFn: () => toggleLike(postId!),
+    mutationFn: () => toggleLike({postId: postId!, token : token! }),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['postDetail', postId] });
+    },
+    onError: () => {
+      alert('추천은 로그인한 사용자만 가능합니다.');
     },
   });
   const isLiking = likeStatus === 'pending';
@@ -61,14 +73,16 @@ const PostDetail = () => {
             <span>조회수 {post.viewCount} </span>
             <span>좋아요 {post.likeCount} </span>
 
-          <button
-              onClick={() => likePost()}
-              disabled={isLiking}
-              className="ml-2 px-3 py-1 bg-[#002561] text-white rounded text-sm disabled:opacity-50"
-          >
-              ❤️ 추천
-          </button>
-        </div>
+            {token && (
+              <button
+                onClick={() => likePost()}
+                disabled={isLiking}
+                className="ml-2 px-3 py-1 bg-[#002561] text-white rounded text-sm disabled:opacity-50"
+              >
+                ❤️ 추천
+              </button>
+            )}
+          </div>
 
           <div className="text-[16px] text-black leading-[28px]">
             {post.content}
