--- src/store/authStore.ts
@@ -1,13 +1,0 @@
-import { create } from 'zustand';
-
-interface AuthState {
-  accessToken: string | null;
-  setAccessToken: (token: string) => void;
-}
-
-export const useAuthStore = create<AuthState>(
-  (set: (partial: Partial<AuthState>) => void) => ({
-    accessToken: null,
-    setAccessToken: (token: string) => set({ accessToken: token }),
-  }),
-);
